+++ model/__init__.py
@@ -1,12 +1,6 @@
 """ 간편한 클래스 임포트
-
 외부 모듈에서 해당 모듈내에 정의된 클래스들을 쉽게 임포트할 수 있도록 클래스들을 임포트해준다.
-
 """
-
-
-from .sample_user_dao         import SampleUserDao
-
 from .admin.event_dao         import EventDao
 from .admin.order_dao         import OrderDao, OrderDetailDao
 from .admin.enquiry_dao       import EnquiryDao
@@ -14,7 +8,6 @@ from .admin.enquiry_dao       import EnquiryDao
 from .admin.seller_dao         import SellerDao
 from .admin.seller_info_dao    import SellerInfoDao
 
-from .admin.product_create_dao import ProductCreateDao
 from .admin.product_manage_dao import ProductManageDao
 from .admin.event_dao         import EventDao
 from .admin.order_dao         import OrderDao, OrderDetailDao
