+++ service/__init__.py
@@ -1,12 +1,6 @@
 """ 간편한 클래스 임포트
-
 외부 모듈에서 해당 모듈내에 정의된 클래스들을 쉽게 임포트할 수 있도록 클래스들을 임포트해준다.
-
 """
-
-
-from .sample_user_service          import SampleUserService
-
 from .admin.order_service          import OrderService
 from .admin.event_service          import EventService
 from .admin.enquiry_service        import EnquiryService
@@ -14,8 +8,7 @@ from .admin.enquiry_service        import EnquiryService
 from .admin.seller_service         import SellerService
 from .admin.seller_info_service    import SellerInfoService
 
-from .admin.product_create_service import ProductCreateService
-from .admin.product_manage_service import ProductManageService
+from .admin.product_service import ProductManageService
 
 from .store.user_service import UserService
 from .store.product_list_service import ProductListService
