+++ utils/connection.py
@@ -7,9 +7,13 @@ import pymysql
 
 
 def get_connection(database):
-    connection = pymysql.connect(host=database['host'],
-                                 user=database['user'],
-                                 password=database['password'],
-                                 db=database['name'],
-                                 charset=database['charset'])
+    
+    connection = pymysql.connect(
+        host     = database['host'],
+        user     = database['user'],
+        password = database['password'],
+        db       = database['name'],
+        charset  = database['charset']
+    )
+    
     return connection
