+++ utils/const.py
@@ -1,0 +1,12 @@
+""" CONST Variable
+
+    Author: 심원두
+
+    History:
+        2021-01-15(심원두): refactoring 초기 작성.
+"""
+
+# 계정 권한 정의
+ACCOUNT_ADMIN  = 1
+ACCOUNT_SELLER = 2
+ACCOUNT_USER   = 3
